+++ lib/vgs.h
@@ -25,6 +25,7 @@
 #pragma once
 #include "vgs_bgm.h"
 #include "vgs_bmpfont.h"
+#include "vgs_calendar.h"
 #include "vgs_cg.h"
 #include "vgs_ctype.h"
 #include "vgs_io.h"
@@ -36,3 +37,13 @@
 #include "vgs_stdlib.h"
 #include "vgs_string.h"
 #include "vgs_system.h"
+
+#ifdef __INTELLISENSE__
+#define LAMBDA(rettype, ARG_LIST, BODY) (rettype(*) ARG_LIST)0
+#else
+#define LAMBDA(rettype, ARG_LIST, BODY)               \
+    ({                                                \
+        rettype __lambda_funcion__ ARG_LIST { BODY; } \
+        __lambda_funcion__;                           \
+    })
+#endif