+++ src/vgsx.h
@@ -27,6 +27,7 @@
 #include <stdint.h>
 #include <string.h>
 #include <math.h>
+#include <time.h>
 #include <functional>
 #include "vdp.hpp"
 
@@ -159,8 +160,8 @@ class VGSX
     void tick();
     void tickSound(int16_t* buf, int samples);
     inline uint32_t* getDisplay() { return this->vdp.ctx.display; }
-    inline int getDisplayWidth() { return VDP_WIDTH; }
-    inline int getDisplayHeight() { return VDP_HEIGHT; }
+    inline int getDisplayWidth() { return VDP_DISPLAY_WIDTH; }
+    inline int getDisplayHeight() { return VDP_DISPLAY_HEIGHT; }
     uint32_t inPort(uint32_t address);
     void outPort(uint32_t address, uint32_t value);
     inline bool isExit() { return this->exitFlag; }
@@ -177,6 +178,12 @@ class VGSX
     void subscribeOutput(std::function<void(uint32_t port, uint32_t value)> callback);
 
   private:
+    static struct tm* now()
+    {
+        time_t now = time(nullptr);
+        return gmtime(&now);
+    }
+
     bool subscribedInput;
     bool subscribedOutput;
     std::function<uint32_t(uint32_t port)> inputCallback;