+++ CHANGES.md
@@ -1,5 +1,18 @@
 # Changes
 
+## Version 0.24.0 **(Disruptive)**
+
+- Support 4k Display
+- Added Calendar API
+- Added Software Reset API
+- Added `slx` and `sly` (Scale Lock X/Y) to OAM.
+- Added `pri` (High Priority Flag) to OAM.
+- Added `vgs_sprite_alpha8` function.
+- Added `vgs_abort` function.
+- Changed the behavior when `OAM.alpha` is set to `0` from “alpha disabled” to “fully transparent.” **(Disruptive)**
+- Changed the behavior when `OAM.scale` is set to `0` from “scaling disabled” to “scaled to 0%.” **(Disruptive)**
+- Changed the default values when initializing a sprite with `vgs_sprite` to `OAM.alpha = 0xFFFFFF` and `OAM.scale = 100`.
+
 ## Version 0.23.0
 
 - Added a Japanese Manual (README-jp.md)